decoder: decouple extension schemas
